Упрощает функции, приводит в порядок названия, решает проблему с генерацией одной карточки
