+++ js/utils/utils.js
@@ -1,6 +1,6 @@
 import {getRandomInteger} from './get-random-positive-integer.js';
 import {getRandomFloatNumber} from './get-random-positive-float.js';
-import {declOfNum} from './decline-numerals.js';
-import {activateElementsInArray} from './activate-elements-in-array.js';
+import {declOfNum} from './declension-of-numerals.js';
+import {activateForm} from './activate-form.js';
 
-export {getRandomInteger, getRandomFloatNumber, declOfNum, activateElementsInArray};
+export {getRandomInteger, getRandomFloatNumber, declOfNum, activateForm};
